split up tests to multiple files
